requirement.txt
streamlit
scikit-learn
joblib
pandas
numpy
newspaper3k
beautifulsoup4
requests
